move http_call into a class of its own
