+++ CKAN-Browser/Makefile
@@ -43,6 +43,7 @@ PY_FILES = \
 	ckan_browser_dialog.py \
 	ckan_browser_dialog_settings.py \
 	ckan_browser_dialog_disclaimer.py \
+	httpcall.py \
 	settings.py \
 	util.py
