+++ CKAN-Browser/ckanconnector.py
@@ -1,37 +1,17 @@
 # -*- coding: utf-8 -*-
 
-import inspect
 import json
 import os
 import sys
 import string
 
-from PyQt5.QtCore import QUrl
-from PyQt5.QtCore import QEventLoop
-from qgis.core import QgsApplication
-from qgis.core import QgsNetworkAccessManager
 
-from urllib.parse import unquote
+from .httpcall import HttpCall
+from .httpcall import RequestsException
+from .httpcall import RequestsExceptionTimeout
+from .httpcall import RequestsExceptionConnectionError
+from .httpcall import RequestsExceptionUserAbort
 from .pyperclip import copy
-
-
-from PyQt5.QtNetwork import *
-
-
-class RequestsException(Exception):
-    pass
-
-
-class RequestsExceptionTimeout(RequestsException):
-    pass
-
-
-class RequestsExceptionConnectionError(RequestsException):
-    pass
-
-
-class RequestsExceptionUserAbort(RequestsException):
-    pass
 
 
 class CkanConnector:
@@ -47,7 +27,6 @@ class CkanConnector:
         # self.sort = 'name asc, title asc'
         self.sort = 'name asc'
         self.mb_downloaded = 0
-        self.reply = None
         self.ua_chrome = {
             b'Accept': b'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
             # DON'T use: haven't found a way to tell QNetworkRequest to decompress on the fly
@@ -71,7 +50,7 @@ class CkanConnector:
         ok, result = self._validate_ckan_url(test_path)
 
         if not ok:
-            return ok,result
+            return ok, result
 
         return self.__get_data(result, 'action/group_list?all_fields=true')
 
@@ -114,7 +93,7 @@ class CkanConnector:
         ok, result = self._validate_ckan_url(self.api)
 
         if not ok:
-            return ok,result
+            return ok, result
 
         self.util.msg_log_debug(group_name)
         if page is None:
@@ -140,7 +119,8 @@ class CkanConnector:
         self.util.msg_log_debug(u'Requesting HEAD for: {0}'.format(url))
 
         try:
-            request_head = self._http_call(url, http_method='head')
+            http_call = HttpCall(self.settings, self.util)
+            request_head = http_call.execute_request(url, http_method='head')
 
             self.util.msg_log_debug(
                 u'get_file_size response:\nex:{0}\nhdr:{1}\nok:{2}\nreason:{3}\nstcode:{4}\nstmsg:{5}\ncontent:{6}'
@@ -160,8 +140,8 @@ class CkanConnector:
                 return False, self.util.tr(u'cc_url_error'), the_exception
 
         except RequestsExceptionTimeout as cte:
-            self.util.msg_log(u'{0}\n{1}\n\n\n{2}'.format(cte, dir(cte), cte.message))
-            return False, self.util.tr(u'cc_connection_timeout').format(cte.message), cte
+            self.util.msg_log(u'{0}\n{1}\n\n\n{2}'.format(cte, dir(cte), cte))
+            return False, self.util.tr(u'cc_connection_timeout').format(cte), cte
         except:
             return False, self.util.tr(u'cc_url_error').format(url, sys.exc_info()[1]), Exception(self.util.tr(u'cc_url_error'))
 
@@ -232,7 +212,8 @@ class CkanConnector:
             # urls might have line breaks
             url = self.util.remove_newline(url)
 
-            response = self._http_call(
+            http_call = HttpCall(self.settings, self.util)
+            response = http_call.execute_request(
                 url
                 , headers=self.ua_chrome
                 , verify=False
@@ -324,7 +305,8 @@ class CkanConnector:
         # url = u'{0}{1}'.format(self.api, unicodedata.normalize('NFKD', action))
         try:
             url = self.util.remove_newline(url)
-            response = self._http_call(
+            http_call = HttpCall(self.settings, self.util)
+            response = http_call.execute_request(
                 url
                 , headers=self.ua_chrome
                 , verify=False
@@ -357,10 +339,10 @@ class CkanConnector:
         except UnicodeEncodeError as uee:
             self.util.msg_log_error(u'msg:{0} enc:{1} args:{2} reason:{3}'.format(uee.message, uee.encoding, uee.args, uee.reason))
             return False, self.util.tr(u'cc_api_not_accessible')
-        except:
-            self.util.msg_log_error(u'unexpected error during request: {0}'.format(sys.exc_info()[0]))
-            self.util.msg_log_last_exception()
-            return False, self.util.tr(u'cc_api_not_accessible')
+        #except:
+        #    self.util.msg_log_error(u'unexpected error during request: {0}'.format(sys.exc_info()[0]))
+        #    self.util.msg_log_last_exception()
+        #    return False, self.util.tr(u'cc_api_not_accessible')
 
         if response.status_code != 200:
             return False, self.util.tr(u'cc_server_fault')
@@ -384,220 +366,6 @@ class CkanConnector:
         if result['success'] is False:
             return False, result['error']['message']
         return True, result['result']
-
-    def _http_call(self, url, **kwargs):
-        """
-        Uses QgsNetworkAccessManager and QgsAuthManager.
-        """
-        method = kwargs.get('http_method', 'get')
-
-        headers = kwargs.get('headers', {})
-        # This fixes a weird error with compressed content not being correctly
-        # inflated.
-        # If you set the header on the QNetworkRequest you are basically telling
-        # QNetworkAccessManager "I know what I'm doing, please don't do any content
-        # encoding processing".
-        # See: https://bugs.webkit.org/show_bug.cgi?id=63696#c1
-        try:
-            del headers[b'Accept-Encoding']
-        except KeyError as ke:
-            # only debugging here as after 1st remove it isn't there anymore
-            self.util.msg_log_debug(u'unexpected error deleting request header: {}'.format(ke))
-            pass
-
-        # Avoid double quoting form QUrl
-        url = unquote(url)
-
-        self.util.msg_log_debug(u'http_call request: {} {}'.format(method, url))
-
-        class Response:
-            status_code = 200
-            status_message = 'OK'
-            text = ''
-            ok = True
-            headers = {}
-            reason = ''
-            exception = None
-
-            def iter_content(self, _):
-                return [self.text]
-
-        self.http_call_result = Response()
-        url = self.util.remove_newline(url)
-
-        req = QNetworkRequest()
-        req.setUrl(QUrl(url))
-
-        for k, v in headers.items():
-            self.util.msg_log_debug("%s: %s" % (k, v))
-            try:
-                req.setRawHeader(k, v)
-            except:
-                self.util.msg_log_error(u'FAILED to set header: {} => {}'.format(k, v))
-                self.util.msg_log_last_exception()
-        if self.auth_cfg:
-            self.util.msg_log_debug(u'before updateNetworkRequest')
-            QgsApplication.authManager().updateNetworkRequest(req, self.auth_cfg)
-            self.util.msg_log_debug(u'before updateNetworkRequest')
-
-        if self.reply is not None and self.reply.isRunning():
-            self.reply.close()
-
-        self.util.msg_log_debug(u'getting QgsNetworkAccessManager.instance()')
-        #func = getattr(QgsNetworkAccessManager.instance(), method)
-        #func = QgsNetworkAccessManager().get(req)
-
-
-
-
-        #manager = QNetworkAccessManager()
-        #event = QEventLoop()
-        #response = manager.get(QNetworkRequest(QUrl(url)))
-        #response.downloadProgress.connect(self.download_progress)
-        #response.finished.connect(event.quit)
-        #event.exec()
-        #response_msg = response.readAll()
-        ##response_msg = str(response_msg)
-        #response_msg = str(response_msg, encoding='utf-8')
-        ##response_msg = response_msg.decode('utf-8')
-        #response.deleteLater()
-        #self.util.msg_log_debug(u'response message:\n{} ...'.format(response_msg[:255]))
-        #self.http_call_result.text = response_msg  # in Python3 all strings are unicode, so QString is not defined
-        #return self.http_call_result
-
-
-
-        # Calling the server ...
-        self.util.msg_log_debug('before self.reply = func(req)')
-        #self.reply = func(req)
-        #self.reply = QgsNetworkAccessManager.instance().get(req)
-        method_call = getattr(QgsNetworkAccessManager.instance(), method)
-        self.reply = method_call(req)
-        #self.reply.setReadBufferSize(1024*1024*1024)
-        #self.reply.setReadBufferSize(1024 * 1024 * 1024 * 1024)
-        self.reply.setReadBufferSize(0)
-        self.util.msg_log_debug('after self.reply = func(req)')
-
-        # Let's log the whole call for debugging purposes:
-        if self.settings.debug:
-            self.util.msg_log_debug("\nSending %s request to %s" % (method.upper(), req.url().toString()))
-            headers = {str(h): str(req.rawHeader(h)) for h in req.rawHeaderList()}
-            for k, v in headers.items():
-                try:
-                    self.util.msg_log_debug("%s: %s" % (k, v))
-                except:
-                    self.util.msg_log_debug('error logging headers')
-
-        if self.auth_cfg:
-            self.util.msg_log_debug("update reply w/ authcfg: {0}".format(self.auth_cfg))
-            QgsApplication.authManager().updateNetworkReply(self.reply, self.auth_cfg)
-
-        self.util.msg_log_debug('before connecting to events')
-
-        # connect downloadProgress event
-        try:
-            self.reply.downloadProgress.connect(self.download_progress)
-            #pass
-        except:
-            self.util.msg_log_error('error connecting "downloadProgress" event')
-            self.util.msg_log_last_exception()
-
-        # connect reply finished event
-        try:
-            self.reply.finished.connect(self.reply_finished)
-            #pass
-        except:
-            self.util.msg_log_error('error connecting reply "finished" progress event')
-            self.util.msg_log_last_exception()
-        self.util.msg_log_debug('after connecting to events')
-
-        # Call and block
-        self.event_loop = QEventLoop()
-        try:
-            self.reply.finished.connect(self.event_loop.quit)
-        except:
-            self.util.msg_log_error('error connecting reply "finished" progress event to event loop quit')
-            self.util.msg_log_last_exception()
-
-        self.mb_downloaded = 0
-        # Catch all exceptions (and clean up requests)
-        self.event_loop.exec()
-
-        # Let's log the whole response for debugging purposes:
-        if self.settings.debug:
-            self.util.msg_log_debug(
-                u'\nGot response [{}/{}] ({} bytes) from:\n{}\nexception:{}'.format(
-                    self.http_call_result.status_code,
-                    self.http_call_result.status_message,
-                    len(self.http_call_result.text),
-                    self.reply.url().toString(),
-                    self.http_call_result.exception
-                )
-            )
-            headers = {str(h): str(self.reply.rawHeader(h)) for h in self.reply.rawHeaderList()}
-            for k, v in headers.items():
-                self.util.msg_log_debug("%s: %s" % (k, v))
-            self.util.msg_log_debug("Payload :\n%s ......" % self.http_call_result.text[:255])
-
-        self.reply.close()
-        self.util.msg_log_debug("Deleting reply ...")
-        try:
-            self.reply.deleteLater()
-        except:
-            self.util.msg_log_error('unexpected error deleting QNetworkReply')
-            self.util.msg_log_last_exception()
-
-        self.reply = None
-
-        if self.http_call_result.exception is not None:
-            self.util.msg_log_error('http_call_result.exception is not None')
-            self.http_call_result.ok = False
-            # raise self.http_call_result.exception
-        return self.http_call_result
-
-    def download_progress(self, bytes_received, bytes_total):
-        mb_received = bytes_received / (1024 * 1024)
-        if mb_received - self.mb_downloaded >= 1:
-            self.mb_downloaded = mb_received
-            self.util.msg_log_debug(
-                u'downloadProgress {:.1f} of {:.1f} MB" '
-                .format(mb_received, bytes_total / (1024 * 1024))
-            )
-
-    def reply_finished(self):
-        self.util.msg_log_debug('------- reply_finished')
-        try:
-            err = self.reply.error()
-            httpStatus = self.reply.attribute(QNetworkRequest.HttpStatusCodeAttribute)
-            httpStatusMessage = self.reply.attribute(QNetworkRequest.HttpReasonPhraseAttribute)
-            self.http_call_result.status_code = httpStatus
-            self.http_call_result.status_message = httpStatusMessage
-            for k, v in self.reply.rawHeaderPairs():
-                self.http_call_result.headers[str(k)] = str(v)
-                self.http_call_result.headers[str(k).lower()] = str(v)
-            if err == QNetworkReply.NoError:
-                self.util.msg_log_debug('QNetworkReply.NoError')
-                self.http_call_result.text = self.reply.readAll()
-                self.http_call_result.ok = True
-            else:
-                self.util.msg_log_error('QNetworkReply Error')
-                self.http_call_result.ok = False
-                msg = "Network error #{0}: {1}"\
-                    .format(
-                        self.reply.error(),
-                        self.reply.errorString()
-                )
-                self.http_call_result.reason = msg
-                self.util.msg_log_error(msg)
-                if err == QNetworkReply.TimeoutError:
-                    self.http_call_result.exception = RequestsExceptionTimeout(msg)
-                if err == QNetworkReply.ConnectionRefusedError:
-                    self.http_call_result.exception = RequestsExceptionConnectionError(msg)
-                else:
-                    self.http_call_result.exception = Exception(msg)
-        except:
-            self.util.msg_log_error(u'unexpected error in {}'.format(inspect.stack()[0][3]))
-            self.util.msg_log_last_exception()
 
     def __get_start(self, page):
         start = self.limit * page - self.limit
